some more refactoring to try to avoid infinite import loop
