+++ lib9ml/python/nineml/user_layer/random.py
@@ -1,4 +1,4 @@
-from .components import BaseComponent
+from .components.base import BaseComponent
 
 
 class RandomDistribution(BaseComponent):
